requirements.txt
streamlit==1.28.0
PyPDF2==3.0.1
openai==0.28.0
gspread==5.9.0
oauth2client==4.1.3
python-dotenv==1.0.0